Pattern and Color fetch created
